Added initial version of dashboard page
The dashboad page is intended to display a live feed of the guesses and perhaps do a confetti explosion when the correct name has been guessed. If I have time it would be nice to have some auto updating statistics, like which names are guessed the most often, which players guess the most, etc.

Currently the page is hooked up to a Fauna stream of all guesses. I started out by doing a document stream but it turns out I would have to query each document as it arrives. By streaming an index I can tap into the values I care about directly from the stream (player and guess).