Fixed timeout issues when making guesses
Ran into a strange issue. Up until now things just worked. However when testing later guesses seemed to not register some times. At other times they would take a very long time to complete.

Not sure this is a proper fix but it seems to work for now. The first step is to actually wait for the Fauna query to finish before returning from the serverless function. Additionally I now explicitly close the Fauna client. It should not be needed because the client is created inside the handler. However after a bit of digging I came across some bugs regarding Node and streams. Given that the connection has always worked reliably when developing locally but not when run in Vercel's serverless functions I suspect that the client might linger around in some weird thawed state. Just in case I close the Fauna client.