Refactor: add all existing guesses at once instead of one by one, etc
Simplified the bootstrapping logic a bit. Removed the document as a stream option. It is not being used. Before each guess would be added individually as state. This caused a lot of re-renders that were not needed. Now we simply pass on the entire list of existing guesses. Reversed the order of guesses to match the streaming behaviour. The last guess should be last.

Refactored, simplified and fixed the behaviour when a new guess is streamed in. Removed listening in on the "start" Fauna stream event type. It was just a number at what time the stream started, not of use to use right now. As such there is no longer any need to check specifically for stream start events. Now once a player has correctly guessed the name the game will stop. The database will still accept new guesses into the database but the dashboard itself will stop updating with new data.